--- src/main/java/com/kyon/todolist/DbRepository.java
@@ -1,7 +1,0 @@
-package com.kyon.todolist;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface DbRepository extends JpaRepository<DbTodo, Integer> { }
